We can now limit the list of studies to a particular user.
Created a new route /user_studies in __init__.py that takes a uva_id.
Pulled out some dupllcate code into a new method; render_study_template

In index.html, added a pulldown to select the user. I pull the list of users from the studies so we only show users that have studies.
Added a script to process the selection.